+++ api/database/migrations/2021_04_02_104755_create_order_items_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     {
         Schema::create('order_items', function (Blueprint $table) {
             $table->id();
-            $table->float('price', unsigned: true);
+            $table->decimal('price', unsigned: true);
             $table->smallInteger('quantity', unsigned: true);
 
             $table->unique(['order_id', 'product_id']);
